1.2.25
See changenotes

+++ admin/manage-invitations/remove-invitation/rmi.php
@@ -1,24 +1,24 @@
 <?php
-	include_once '../../includes/db_connect.php';
-	include_once '../../includes/functions.php';
-	secure_session_start();
+	require('constants.php');
+	require_once(ABS_PATH.INC_PATH.'functions.php');
 
+	secure_session_start();
 	if(isset($_POST['uid'])){
-		if(login_check($mysqli)){
+		if(login_check($mysqli) && $_SESSION['user']['is_admin']){
 			if(isUser($_POST['uid'], $mysqli)){
 				$uid = $_POST['uid'];
 				$mysqli->query("DELETE FROM ausstehende_einladungen WHERE uid = '$uid'");
-				unlink("../../../users/$uid/$uid.json");
-				unlink("../../../users/$uid/avatar.jpg");
-				rmdir("../../../users/$uid/");
-				echo success(["message"=>"Einladung wurde entfernt"]);
+				unlink(ABS_PATH . "/users/$uid/$uid.json");
+				unlink(ABS_PATH . "/users/$uid/avatar.jpg");
+				rmdir(ABS_PATH . "/users/$uid/");
+				success(["message"=>"Einladung wurde entfernt"]);
 			} else {
-				echo error('internalError', 500, 'Cannot find user in database');
+				error('internalError', 500, 'Cannot find user in database');
 			}
 		} else {
-			echo error('clientError', 403, 'Forbidden');
+			error('clientError', 403, 'Forbidden');
 		}
 	} else {
-		echo error('clientError', 400, 'Bad Request');
+		error('clientError', 400, 'Bad Request');
 	}
 ?>
